[ALPS07158507] GPU: m1.13 k4.19 fix Google security CVE-2021-39661
[Detail]
CVE-2021-39661
Ticket:132620

[Solution]
PP132620.CL6189956.CVE-2021-39661.Fix.not.validated.input.
arguments.to.MMU_MapPages.patch

Fix for not validated input arguments to MMU_MapPages()
The number of the pages was not validated in MMU_MapPages().
This change adds validation of this parameter agains the allocated pages
count in the PMR on the level of PMR factory.

MTK-Commit-Id: 6c8450e34499866b1c077b9b2829e1172804a0d3

CR-Id: ALPS07158507
Feature: IMG GPU DDK driver
Signed-off-by: “Yanqiong <yanqiong.tao@mediatek.com>
Change-Id: I23a0e67edac559ead34bd4a3979a96f0fa865cf0