[ALPS07360767] xhci: clear reserved field of endpoint context
We use reserved field of endpoint context to schedule periodic
transfer. We just used or | = to recover reserved field
of endpoint context, before. So now we clear reserved field of
endpoint context, when we drop enpoint. This ensures that the
current offset is used to schedule periodic endpoint.

MTK-Commit-Id: 5bdc2c9a72640623f362ef3f17508052a7161d0f

Signed-off-by: ot903067 <ot_he.liu@mediatek.com>
CR-Id: ALPS07360767
Feature: USB Audio Middleware
Change-Id: I13bf1773c279a2b91c15b07e42f071c79be334b6